Added failing test for field renames not renaming indexes. Refs #23577
The ticket is regarding migrations that rename a field or model. They do
not also rename associated indexes, and thus you will get an error if
you later add a new field with the same name that has an index.

This commit adds a test for index names after renaming an indexed field.